Split long files to smaller ones
