+++ src/core/config/version.h
@@ -1,0 +1,35 @@
+// Copyright 2025 Timo Heimonen <timo.heimonen@proton.me>
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+//
+/**
+ * @file version.h
+ * @brief Software version information
+ * @author Timo Heimonen <timo.heimonen@proton.me>
+ * @date 2025
+ *
+ * This header defines the software version number.
+ */
+#ifndef VERSION_H
+#define VERSION_H
+
+// --- Version Information ---
+/**
+ * @def SOFTVERSION
+ * @brief Software version number
+ */
+#define SOFTVERSION 0.51f
+
+#endif // VERSION_H
+
