+++ src/output/console/statistics.cpp
@@ -21,6 +21,7 @@
 #include <vector>      // Required for std::vector
 #include "core/config/constants.h" // Constants for precision values
 #include "output/console/messages.h"  // Centralized messages
+#include "output/console/statistics.h"  // Function declaration
 
 // --- Helper structures and functions for statistics ---
